# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI SkillFit API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin officer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInfo"
        "401":
          description: Not authenticated
  /admin/questions-pool:
    get:
      operationId: getAdminQuestionsPool
      tags: [admin]
      summary: "[Admin] Large slice of question pool for QA / rubric review"
      parameters:
        - name: trade
          in: query
          required: true
          schema:
            type: string
        - name: language
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        "200":
          description: Question pool entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "401":
          description: Not authenticated
  /admin/actions:
    post:
      operationId: createOfficerAction
      tags: [admin]
      summary: Record an officer action on a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficerActionBody"
      responses:
        "201":
          description: Action recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficerAction"
  /admin/actions/{candidateId}:
    get:
      operationId: getCandidateActions
      tags: [admin]
      summary: Get audit log for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of officer actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfficerAction"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Question:
      type: object
      properties:
        id:
          type: string
        trade:
          type: string
        language:
          type: string
        text:
          type: string
        category:
          type: string
          description: Rubric bucket (experience, safety, skill, etc.)
        difficulty:
          type: string
          enum: [easy, medium, hard]
          description: Adaptive difficulty level for the current question
        audioHint:
          type: string
          nullable: true
      required:
        - id
        - trade
        - language
        - text
...
    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        username:
          type: string
      required:
        - success
        - username
    AdminInfo:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
      required:
        - username
        - role
    OfficerActionBody:
      type: object
      properties:
        candidateId:
          type: integer
        interviewId:
          type: integer
          nullable: true
        action:
          type: string
          enum: [shortlist, send_to_training, request_reinterview, escalate, approve, reject]
        notes:
          type: string
          nullable: true
      required:
        - candidateId
        - action
    OfficerAction:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        interviewId:
          type: integer
          nullable: true
        action:
          type: string
        officerUsername:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - candidateId
        - action
        - officerUsername
        - createdAt
...